+++ moyo/src/lib.rs
@@ -86,6 +86,7 @@ pub mod data;
 pub mod identify;
 pub mod math;
 pub mod search; // Public for benchmarking
+pub mod subgroup;
 pub mod utils;
 
 mod symmetrize;